exceptions: Dump registers as 64-bit values

+++ source/core/exceptions.cpp
@@ -16,45 +16,45 @@ static void Print_Registers(struct AArch64_Regs* Regs)
 	Log() << fmt::endl << "AArch64 general purpose registers state:" << fmt::endl;
 
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x0  = 0x" << (int32_t)Regs->x0 << "  x1  = 0x" << (int32_t)Regs->x1
-	      <<  "  x2  = 0x" << (int32_t)Regs->x2 << "  x3  = 0x" << (int32_t)Regs->x3
+	      <<  "  x0  = 0x" << Regs->x0 << "  x1  = 0x" << Regs->x1
+	      <<  "  x2  = 0x" << Regs->x2 << "  x3  = 0x" << Regs->x3
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x4  = 0x" << (int32_t)Regs->x4 << "  x5  = 0x" << (int32_t)Regs->x5
-	      <<  "  x6  = 0x" << (int32_t)Regs->x6 << "  x7  = 0x" << (int32_t)Regs->x7
+	      <<  "  x4  = 0x" << Regs->x4 << "  x5  = 0x" << Regs->x5
+	      <<  "  x6  = 0x" << Regs->x6 << "  x7  = 0x" << Regs->x7
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x8  = 0x" << (int32_t)Regs->x8 <<  "  x9  = 0x" << (int32_t)Regs->x9
-	      <<  "  x10 = 0x" << (int32_t)Regs->x10 << "  x11 = 0x" << (int32_t)Regs->x11
+	      <<  "  x8  = 0x" << Regs->x8 <<  "  x9  = 0x" << Regs->x9
+	      <<  "  x10 = 0x" << Regs->x10 << "  x11 = 0x" << Regs->x11
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x12 = 0x" << (int32_t)Regs->x12 << "  x13 = 0x" << (int32_t)Regs->x13
-	      <<  "  x14 = 0x" << (int32_t)Regs->x14 << "  x15 = 0x" << (int32_t)Regs->x15
+	      <<  "  x12 = 0x" << Regs->x12 << "  x13 = 0x" << Regs->x13
+	      <<  "  x14 = 0x" << Regs->x14 << "  x15 = 0x" << Regs->x15
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x16 = 0x" << (int32_t)Regs->x16 << "  x17 = 0x" << (int32_t)Regs->x17
-	      <<  "  x18 = 0x" << (int32_t)Regs->x18 << "  x19 = 0x" << (int32_t)Regs->x19
+	      <<  "  x16 = 0x" << Regs->x16 << "  x17 = 0x" << Regs->x17
+	      <<  "  x18 = 0x" << Regs->x18 << "  x19 = 0x" << Regs->x19
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x20 = 0x" << (int32_t)Regs->x20 << "  x21 = 0x" << (int32_t)Regs->x21
-	      <<  "  x22 = 0x" << (int32_t)Regs->x22 << "  x23 = 0x" << (int32_t)Regs->x23
+	      <<  "  x20 = 0x" << Regs->x20 << "  x21 = 0x" << Regs->x21
+	      <<  "  x22 = 0x" << Regs->x22 << "  x23 = 0x" << Regs->x23
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x24 = 0x" << (int32_t)Regs->x24 << "  x25 = 0x" << (int32_t)Regs->x25
-	      <<  "  x26 = 0x" << (int32_t)Regs->x26 << "  x27 = 0x" << (int32_t)Regs->x27
+	      <<  "  x24 = 0x" << Regs->x24 << "  x25 = 0x" << Regs->x25
+	      <<  "  x26 = 0x" << Regs->x26 << "  x27 = 0x" << Regs->x27
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  x28 = 0x" << (int32_t)Regs->x28 << "  x29 = 0x" << (int32_t)Regs->x29
-	      <<  "  lr =  0x" << (int32_t)Regs->lr <<  "  sp =  0x" << (int32_t)Regs->sp
+	      <<  "  x28 = 0x" << Regs->x28 << "  x29 = 0x" << Regs->x29
+	      <<  "  lr =  0x" << Regs->lr <<  "  sp =  0x" << Regs->sp
 	      << fmt::endl;
 	Log() << fmt::fill << fmt::hex
-	      <<  "  pc  = 0x" << (int32_t)Regs->pc
+	      <<  "  pc  = 0x" << Regs->pc
 	      << fmt::endl;
 
 	Log() << fmt::endl << "AArch64 control registers state:" << fmt::endl;
 
 	Log() << fmt::fill << fmt::hex
-	      <<  "  cpsr = 0x" << (int32_t)Regs->cpsr
+	      <<  "  cpsr = 0x" << Regs->cpsr
 	      << fmt::endl;
 }
